Spring Boot

+++ src/main/java/com/project/gym/repository/Berlet_vasarlasRep.java
@@ -4,5 +4,5 @@ import com.project.gym.model.Berlet_vasarlas;
 import com.project.gym.model.Felhasznalo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface Berlet_vasarlasRep extends JpaRepository<Berlet_vasarlas,Integer> {
+public interface Berlet_vasarlasRep extends JpaRepository<Berlet_vasarlas,Long> {
 }
