+++ src/main/java/com/project/gym/repository/CegRep.java
@@ -4,5 +4,5 @@ import com.project.gym.model.Ceg;
 import com.project.gym.model.Felhasznalo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CegRep extends JpaRepository<Ceg,Integer> {
+public interface CegRep extends JpaRepository<Ceg,Long> {
 }
