+++ src/main/java/com/project/gym/repository/RendelesRep.java
@@ -4,5 +4,5 @@ import com.project.gym.model.Felhasznalo;
 import com.project.gym.model.Rendeles;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RendelesRep extends JpaRepository<Rendeles,Integer> {
+public interface RendelesRep extends JpaRepository<Rendeles,Long> {
 }
