+++ src/main/java/com/project/gym/repository/Szemelyi_edzoRep.java
@@ -4,5 +4,5 @@ import com.project.gym.model.Felhasznalo;
 import com.project.gym.model.Szemelyi_edzo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface Szemelyi_edzoRep extends JpaRepository<Szemelyi_edzo,Integer> {
+public interface Szemelyi_edzoRep extends JpaRepository<Szemelyi_edzo,Long> {
 }
