+++ src/main/java/com/project/gym/repository/TermekRep.java
@@ -4,5 +4,5 @@ import com.project.gym.model.Felhasznalo;
 import com.project.gym.model.Termek;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TermekRep extends JpaRepository<Termek,Integer> {
+public interface TermekRep extends JpaRepository<Termek,Long> {
 }
